requirements.txt
langchain-community>=0.0.10
python-dotenv>=0.19.0
streamlit>=1.18.0
google-generativeai>=0.3.0
requests>=2.31.0
pillow>=10.0.0
opencv-python>=4.8.0
azure-cognitiveservices-vision-computervision>=0.9.0
msrest>=0.7.1
SpeechRecognition>=3.10.0
PyMuPDF>=1.23.0
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
folium>=0.14.0
streamlit-folium>=0.15.0
geopy>=2.4.0
google>=3.0.0
pytesseract>=0.3.10
